+++ CHANGES.md
@@ -2,6 +2,11 @@
 
 ## 1.1.0 (not yet released)
 
+- Add `manta.CLISigner` for signing appropriate for an CLI tool. It will
+  use an ssh-agent if available and fallback to using a given SSH key
+  (looking at key files in "~/.ssh/" as appropriate). Switch `mantash`
+  to use this for signing.
+
 - Fix a bug where the '-type' argument to 'mantash find' always returned empty
   results.